pyproject.toml
[project]
name = "message-blaster"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django>=5.1.3",
    "celery>=5.4.0",
    "python-telegram-bot>=21.7",
    "djangorestframework>=3.15.2",
    "valkey>=6.0.2",
    "psycopg2-binary>=2.9.10",
    "django-model-utils>=5.0.0",
    "telethon>=1.38.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pytest>=8.3.3",
    "factory-boy>=3.3.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/message_blaster"]
